refactor: move assert to be internal focussed, improve errors thrown
