+++ std/_util/deep_assign.ts
@@ -1,5 +1,5 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import { assert } from "../validation/assert.ts";
+import { assert } from "../_util/assert.ts";
 
 export function deepAssign(
   target: Record<string, unknown>,
