+++ std/archive/tar.ts
@@ -28,7 +28,7 @@
  */
 import { MultiReader } from "../io/readers.ts";
 import { BufReader } from "../io/bufio.ts";
-import { assert } from "../validation/assert.ts";
+import { assert } from "../_util/assert.ts";
 
 const recordSize = 512;
 const ustar = "ustar\u000000";
