+++ std/encoding/csv.ts
@@ -7,7 +7,7 @@
 import { BufReader } from "../io/bufio.ts";
 import { TextProtoReader } from "../textproto/mod.ts";
 import { StringReader } from "../io/readers.ts";
-import { assert } from "../validation/assert.ts";
+import { assert } from "../_util/assert.ts";
 
 const INVALID_RUNE = ["\r", "\n", '"'];
