+++ std/flags/mod.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 
-import { assert } from "../validation/assert.ts";
+import { assert } from "../_util/assert.ts";
 
 export interface Args {
   /** Contains all the arguments that didn't have an option associated with
