+++ std/fs/copy.ts
@@ -2,7 +2,7 @@
 import * as path from "../path/mod.ts";
 import { ensureDir, ensureDirSync } from "./ensure_dir.ts";
 import { isSubdir, getFileInfoType } from "./_util.ts";
-import { assert } from "../validation/assert.ts";
+import { assert } from "../_util/assert.ts";
 
 const isWindows = Deno.build.os === "windows";
