+++ std/log/mod.ts
@@ -7,7 +7,7 @@ import {
   FileHandler,
   RotatingFileHandler,
 } from "./handlers.ts";
-import { assert } from "../validation/assert.ts";
+import { assert } from "../_util/assert.ts";
 import { LevelName } from "./levels.ts";
 
 export { LogLevels } from "./levels.ts";
