+++ std/uuid/v5.ts
@@ -8,7 +8,7 @@ import {
 } from "./_common.ts";
 import { Sha1 } from "../hash/sha1.ts";
 import { isString } from "../node/util.ts";
-import { assert } from "../validation/assert.ts";
+import { assert } from "../_util/assert.ts";
 
 const UUID_RE = /^[0-9a-f]{8}-[0-9a-f]{4}-[5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
